struct completer now provide snippets
