+++ main/src/ca/uwaterloo/flix/api/lsp/provider/completion/TypeCompleter.scala
@@ -39,28 +39,4 @@ object TypeCompleter {
     else
       Priority.Lower
   }
-
-  /**
-    * Format type params in the right form to be inserted as a snippet
-    * e.g. "[${1:a}, ${2:b}, ${3:c}]"
-    */
-  def formatTParamsSnippet(tparams: List[TypedAst.TypeParam]): String = {
-    tparams match {
-      case Nil => ""
-      case _ => tparams.zipWithIndex.map {
-        case (tparam, idx) => "$" + s"{${idx + 1}:${tparam.name}}"
-      }.mkString("[", ", ", "]")
-    }
-  }
-
-  /**
-    * Format type params in the right form to be displayed in the list of completions
-    * e.g. "[a, b, c]"
-    */
-  def formatTParams(tparams: List[TypedAst.TypeParam]): String = {
-    tparams match {
-      case Nil => ""
-      case _ => tparams.map(_.name).mkString("[", ", ", "]")
-    }
-  }
 }
